% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phosphoproteome.R
\name{nrowToSingleString}
\alias{nrowToSingleString}
\title{Merge some character into one character}
\usage{
nrowToSingleString(x)
}
\arguments{
\item{x}{A data.frame that contains the same one protein accession and different phosphorylated sites}
}
\value{
a vector contains one character in which all the phosphorylated site information were recorded
}
\description{
Merge the information of different phosphorylated sites in one protein into one row in a data.frame
}
