% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phosphoproteome.R
\name{PhosphorylatedProteinInformation}
\alias{PhosphorylatedProteinInformation}
\title{Extract the phosphorylated proteins and the corresponding phosphorylated sites.}
\usage{
PhosphorylatedProteinInformation(x)
}
\arguments{
\item{x}{A data.frame}
}
\value{
A data.frame in whcih the first column is the unique phosphorylated protein accession and the second column is the sum of the phosphorylated sites
}
\description{
Extract the phosphorylated proteins and the corresponding phosphorylated sites from the export .xlsx data created by Proteome Discovery 2.4
Note: only the first protein and the first position will be be remained
}
