% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phosphoproteome.R
\name{sumHighConfidentSites}
\alias{sumHighConfidentSites}
\title{Extract the high confident phosphorylated sites}
\usage{
sumHighConfidentSites(x)
}
\arguments{
\item{x}{the result of PhosphorylatedProteinInformation,namely a data.frame containing one column of unique protein and all of the phosphorylated information.}
}
\value{
a list which consist of data.frame, the name of each data.frame is the unique protein accession, the data.frame contains the separated site and phosphoScore and HighConfident
}
\description{
High confident phosphorylated sites: score > 75
}
