% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phosphoproteome.R
\name{SummaryTotalPhosphosites}
\alias{SummaryTotalPhosphosites}
\title{Extract the high confident phosphorylated sites}
\usage{
SummaryTotalPhosphosites(x)
}
\arguments{
\item{x}{the result of sumHighConfidentSites,namely a list of all phosphorylated proteins。}
}
\value{
a data. frame contains Total phosphoSites, The most phosphorylated protein, The number of phosphoSites of most phosphorylated protein.
}
\description{
The summary of Total phosphoSites and the most phosphorylated protein
}
