% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phosphoproteome.R
\name{tolowerPhosphoSitesinFasta}
\alias{tolowerPhosphoSitesinFasta}
\title{To label the phosphorylated amino acid in protein fasta file}
\usage{
tolowerPhosphoSitesinFasta(x, fasta)
}
\arguments{
\item{x}{a list, the result of sumHighConfidentSites}

\item{fasta}{a fasta dataset download form Uniprot}
}
\value{
a fasta file of identified phosphorylated proteins, which was labeled with phosphorylated sites
}
\description{
To tolower the phosphorylated amino acid in protein fasta file
}
