% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phosphoproteome.R
\name{tolowerSpecificSite}
\alias{tolowerSpecificSite}
\title{Batch tolower specific postion letters}
\usage{
tolowerSpecificSite(x, y)
}
\arguments{
\item{x}{a vector that contains one character}

\item{y}{a numeric vector}
}
\value{
a vector contains one character that will be labeled with lowercases in the specific position
}
\description{
To tolower specific postion letters in a vector that contains one character
}
